requirements.txt
asgiref==3.8.1
channels==4.0.0
channels-redis==4.2.0
crispy-bootstrap5
Django==5.0.7
django-crispy-forms
dj-database-url==2.1.0 
gunicorn==22.0.0
psycopg2-binary
python-decouple==3.8
pytz==2024.1
sqlparse==0.5.0
tzdata==2024.1
whitenoise==6.7.0
python-json-logger==2.0.7
sentry-sdk==1.40.0
colorlog==6.8.0
setuptools>=65.0.0
django-redis>=5.4.0
celery==5.3.4
redis==5.0.1
Pillow
imagehash==4.3.1
django-storages[s3]==1.14.2
boto3==1.34.9
django-storages[google]==1.14.2
google-cloud-storage==2.14.0
pytils>=0.3.0
django-silk==5.4.3
django-q2
daphne==4.1.2
flower












